Tehtävä 4.21*

+++ osanelja/blogilista/controllers/blogs.js
@@ -53,7 +53,30 @@ blogsRouter.post('/', async (request, response) => {
 
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'Blog not in database' })
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  if (!user) {
+    return response.status(400).json({ error: 'UserId missing or not valid' })
+  }
+
+  if ( blog.user.toString() !== decodedToken.id.toString() ) {
+    return response.status(401).json({ error: 'Unauthorized' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
+  await User.findByIdAndUpdate(
+    blog.user,
+    { $pull: { blogs: blog.id } })
+
   response.status(204).end()
 })
